requirements.txt
httpx==0.27.0
feedparser==6.0.11
aiosqlite==0.20.0
python-dateutil==2.9.0
typer==0.12.3
rich==13.7.1